requirements.txt
selenium
requests
pytesseract
pillow
beautifulsoup4
opencv-python
tqdm
pandas
webdriver-manager
python-dotenv